Reset FBNeo runtime to stock Fightcade source

+++ Sources/FightcadeFBNeo/src/intf/video/sdl/vid_sdlfx.cpp
@@ -255,7 +255,7 @@ static int Frame(bool bRedraw)						// bRedraw = 0
 // Paint the BlitFX surface onto the primary surface
 static int Paint(int bValidate)
 {
-	SDL_Rect sdlrDest = { 0, 0, static_cast<Uint16>(nGameWidth * nSize), static_cast<Uint16>(nGameHeight * nSize) };
+	SDL_Rect sdlrDest = { 0, 0, nGameWidth * nSize, nGameHeight * nSize };
 
 	if (bValidate & 2) {
 		MemToSurf();									// Copy the memory buffer to the directdraw buffer for later blitting
